Added some poker scoring algorithm logic
This logic is fuzzy, since there is no real way to determine a tie breaker
yet. The 'score' structure stores the hand type achieved, but it doesn't
store any sort of extra data, like the high cards, etc.

I may be able to solve this by storing a pair of 'meta cards' in the score
objects, which would be extremely context sensitive, and only used in the
case of a tie.